requirements.txt
flask==3.0.0
python-dotenv==1.0.0
hypercorn==0.15.0
asgiref==3.7.2
llama-index-llms-anthropic==0.1.3
llama-index-core==0.10.1